openai_agents_with_fastapi/backend/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
openai-agent